Move trainables to separate file.
